+++ src/main/java/br/gov/mj/sislegis/app/service/ComentarioService.java
@@ -1,13 +1,11 @@
 package br.gov.mj.sislegis.app.service;
 
-import java.math.BigInteger;
-import java.util.List;
-
-import javax.ejb.Local;
-
 import br.gov.mj.sislegis.app.json.ComentarioJSON;
 import br.gov.mj.sislegis.app.model.Comentario;
 import br.gov.mj.sislegis.app.model.Usuario;
+
+import javax.ejb.Local;
+import java.util.List;
 
 @Local
 public interface ComentarioService extends Service<Comentario> {
@@ -16,7 +14,7 @@ public interface ComentarioService extends Service<Comentario> {
 
 	List<ComentarioJSON> findByProposicao(Long idProposicao, Integer posicaoInicial, Integer limite);
 
-	BigInteger totalByProposicao(Long idProposicao);
+	Long totalByProposicao(Long idProposicao);
 
 	void salvarComentario(ComentarioJSON comentarioJSON, Usuario usuario) throws IllegalAccessException;
